requirements.txt
# 核心深度学习库
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.5
tensorboard>=2.14.0

# 图像处理
Pillow>=10.1.0

# 数值计算
numpy>=1.24.0

# 机器学习工具
scikit-learn>=1.3.0

# 进度显示
tqdm>=4.66.0

# 可视化与终端美化
rich>=13.6.0
tabulate>=0.9.0

# 配置与YAML支持
PyYAML>=6.0.1

# GUI组件
customtkinter>=5.2.0

# 模型评估工具（如用于后续评估）
torchmetrics>=0.10.0
